test: move remote account provider lru tests
