+++ magicblock-chainlink/src/remote_account_provider/tests.rs
@@ -658,3 +658,184 @@ async fn test_get_accounts_until_slots_match_finding_matching_slot_but_one_accou
     assert!(accs[0].is_found());
     assert!(!accs[1].is_found());
 }
+
+// -----------------
+// LRU Cache/Eviction/Removal
+// -----------------
+async fn setup_with_accounts(
+    pubkeys: &[Pubkey],
+    accounts_capacity: usize,
+) -> (
+    RemoteAccountProvider<ChainRpcClientMock, ChainPubsubClientMock>,
+    mpsc::Receiver<ForwardedSubscriptionUpdate>,
+    mpsc::Receiver<Pubkey>,
+) {
+    let rpc_client = {
+        let mut rpc_client_builder = ChainRpcClientMockBuilder::new().slot(1);
+        for pubkey in pubkeys {
+            rpc_client_builder = rpc_client_builder.account(
+                *pubkey,
+                Account {
+                    lamports: 555,
+                    data: vec![],
+                    owner: system_program::id(),
+                    executable: false,
+                    rent_epoch: 0,
+                },
+            );
+        }
+        rpc_client_builder.build()
+    };
+
+    let (tx, rx) = mpsc::channel(1);
+    let pubsub_client = ChainPubsubClientMock::new(tx, rx);
+
+    let (forward_tx, forward_rx) = mpsc::channel(100);
+    let (subscribed_accounts, config) =
+        create_test_lru_cache(accounts_capacity);
+    let chain_slot = Arc::<AtomicU64>::default();
+
+    let provider = RemoteAccountProvider::new(
+        rpc_client,
+        pubsub_client,
+        forward_tx,
+        &config,
+        subscribed_accounts,
+        ChainSlot::new(chain_slot),
+    )
+    .await
+    .unwrap();
+
+    let removed_account_tx = provider.try_get_removed_account_rx().unwrap();
+    (provider, forward_rx, removed_account_tx)
+}
+
+fn drain_removed_account_rx(rx: &mut mpsc::Receiver<Pubkey>) -> Vec<Pubkey> {
+    let mut removed_accounts = Vec::new();
+    while let Ok(pubkey) = rx.try_recv() {
+        removed_accounts.push(pubkey);
+    }
+    removed_accounts
+}
+
+#[tokio::test]
+async fn test_add_accounts_up_to_limit_no_eviction() {
+    // Higher level version (including removed_rx) from
+    // src/remote_account_provider/lru_cache.rs:
+    // - test_lru_cache_add_accounts_up_to_limit_no_eviction
+    init_logger();
+
+    let pubkey1 = Pubkey::new_unique();
+    let pubkey2 = Pubkey::new_unique();
+    let pubkey3 = Pubkey::new_unique();
+
+    let pubkeys = &[pubkey1, pubkey2, pubkey3];
+
+    let (provider, _, mut removed_rx) = setup_with_accounts(pubkeys, 3).await;
+
+    // Add three accounts (up to limit)
+    for pk in pubkeys {
+        provider
+            .try_get(*pk, AccountFetchOrigin::GetAccount)
+            .await
+            .unwrap();
+    }
+
+    // No evictions should occur
+    let removed = drain_removed_account_rx(&mut removed_rx);
+    debug!(removed = ?removed, "Removed accounts");
+    assert!(removed.is_empty(), "Expected no removed accounts");
+}
+
+#[tokio::test]
+async fn test_eviction_order() {
+    // Higher level version (including removed_rx) from
+    // src/remote_account_provider/lru_cache.rs:
+    // - test_lru_cache_lru_eviction_order
+    init_logger();
+
+    let pubkey1 = Pubkey::new_unique();
+    let pubkey2 = Pubkey::new_unique();
+    let pubkey3 = Pubkey::new_unique();
+    let pubkey4 = Pubkey::new_unique();
+    let pubkey5 = Pubkey::new_unique();
+
+    let pubkeys = &[pubkey1, pubkey2, pubkey3, pubkey4, pubkey5];
+    let (provider, _, mut removed_rx) = setup_with_accounts(pubkeys, 3).await;
+
+    // Fill cache: [1, 2, 3] (1 is least recently used)
+    provider
+        .try_get(pubkey1, AccountFetchOrigin::GetAccount)
+        .await
+        .unwrap();
+    provider
+        .try_get(pubkey2, AccountFetchOrigin::GetAccount)
+        .await
+        .unwrap();
+    provider
+        .try_get(pubkey3, AccountFetchOrigin::GetAccount)
+        .await
+        .unwrap();
+
+    // Access pubkey1 to make it more recently used: [2, 3, 1]
+    // This should just promote, making order [2, 3, 1]
+    provider
+        .try_get(pubkey1, AccountFetchOrigin::GetAccount)
+        .await
+        .unwrap();
+
+    // Add pubkey4, should evict pubkey2 (now least recently used)
+    provider
+        .try_get(pubkey4, AccountFetchOrigin::GetAccount)
+        .await
+        .unwrap();
+
+    // Check channel received the evicted account
+
+    let removed_accounts = drain_removed_account_rx(&mut removed_rx);
+    assert_eq!(removed_accounts, [pubkey2]);
+
+    // Add pubkey5, should evict pubkey3 (now least recently used)
+    provider
+        .try_get(pubkey5, AccountFetchOrigin::GetAccount)
+        .await
+        .unwrap();
+
+    // Check channel received the second evicted account
+    let removed_accounts = drain_removed_account_rx(&mut removed_rx);
+    assert_eq!(removed_accounts, [pubkey3]);
+}
+
+#[tokio::test]
+async fn test_multiple_evictions_in_sequence() {
+    // Higher level version (including removed_rx) from
+    // src/remote_account_provider/lru_cache.rs:
+    // - test_lru_cache_multiple_evictions_in_sequence
+    init_logger();
+
+    // Create test pubkeys
+    let pubkeys: Vec<Pubkey> = (1..=7).map(|_| Pubkey::new_unique()).collect();
+
+    let (provider, _, mut removed_rx) = setup_with_accounts(&pubkeys, 4).await;
+
+    // Fill cache to capacity (no evictions)
+    for pk in pubkeys.iter().take(4) {
+        provider
+            .try_get(*pk, AccountFetchOrigin::GetAccount)
+            .await
+            .unwrap();
+    }
+
+    // Add more accounts and verify evictions happen in LRU order
+    for i in 4..7 {
+        provider
+            .try_get(pubkeys[i], AccountFetchOrigin::GetAccount)
+            .await
+            .unwrap();
+        let expected_evicted = pubkeys[i - 4]; // Should evict the account added 4 steps ago
+
+        // Verify the evicted account was sent over the channel
+        let removed_accounts = drain_removed_account_rx(&mut removed_rx);
+        assert_eq!(removed_accounts, vec![expected_evicted]);
+    }
+}
